+++ internal/infrastracture/models/blog.go
@@ -65,55 +65,3 @@ func (blogs Blogs) ToSlice() []*Blog {
 	}
 	return result
 }
-
-type UserId int64
-
-type User struct {
-	Id       UserId `json:"id,omitempty" db:"id"`
-	Name     string `json:"name" db:"name"`
-	Email    string `json:"email,omitempty" db:"email"`
-	Password string `json:"password,omitempty" db:"password"`
-	Created  uint   `json:"created,omitempty" db:"created"`
-	Modified uint   `json:"modified,omitempty" db:"modified"`
-}
-
-type TagId int64
-
-type Tag struct {
-	Id   TagId  `json:"id" db:"id"`
-	Name string `json:"name" db:"name"`
-}
-
-type Tags []*Tag
-
-type BlogsTags struct {
-	BlogId BlogId `json:"blogId" db:"blog_id"`
-	TagId  TagId  `json:"tagId" db:"tag_id"`
-	Name   string `json:"name" db:"name"`
-}
-
-// BlogsTagsArray は、ブログとタグのリレーションを保持する構造体のスライス
-type BlogsTagsArray []*BlogsTags
-
-// TagIds は、リレーションのスライスからタグIDのスライスを取得する関数
-func (arr BlogsTagsArray) TagIds() []TagId {
-	var tags []TagId
-	for _, bt := range arr {
-		tags = append(tags, bt.TagId)
-	}
-	return tags
-}
-
-// TagNames は、リレーションのスライスからタグ名のスライスを取得する関数
-func (arr BlogsTagsArray) TagNames() []string {
-	var tags []string
-	for _, t := range arr {
-		tags = append(tags, t.Name)
-	}
-	return tags
-}
-
-// Contains は、リレーションのスライスに指定したタグが含まれているかを判定する関数
-func (arr BlogsTagsArray) Contains(tag string) bool {
-	return slices.Contains(arr.TagNames(), tag)
-}
